+++ src/distarter/main.py
@@ -3,10 +3,11 @@
 this file serves as Python code to get started with domible
 """
 
+import domible
 from domible.elements import Html, Body, Title
 from domible.elements import Heading, Anchor, Paragraph, HorizontalRule 
 from domible.starterDocuments import basic_head_empty_body
-from domible.tools import open_html_in_browser
+from domible.tools import open_html_document_in_browser
 
 
 #######
@@ -19,16 +20,17 @@ def run() -> None:
     cib_anchor = Anchor(
         href="http://codinginblind.vip", contents="the Coding In Blind website"
     )
+    version = domible.__version__
     htmlDoc = basic_head_empty_body(title)
     body = htmlDoc.get_body_element()
     body.add_content([
-        Heading(1, title), 
+        Heading(1, title),
+        Paragraph(f"using domible version {version}"),
         Paragraph(f"check out: {cib_anchor}"),
         HorizontalRule(),
         Paragraph('Adding a non printable character, codepoint = 0xfeff, in the square brackets [\ufeff] to make sure encodings are set correctly.')
     ])
-    open_html_in_browser(htmlDoc)
-    ## open_html_in_browser(htmlDoc, "saved_distarter.html", True)
+    open_html_document_in_browser(htmlDoc)
 
 
 if __name__ == "__main__":